migrate remaining util tests
